requirements.txt
# requirements.txt
rich>=10.0.0
requests>=2.25.1
python-dotenv>=0.19.0
typer>=0.4.0
pydantic>=2.0.0
